frontend/cc_proyecto/requirements.txt
django=3.1.2
djangorestframework=3.12.2
psycopg2-binary=2.9.5